requirements.txt
discord-py-interactions~=4.3.4
undetected-chromedriver~=3.1.6
interactions-tasks~=1.0.0
pystray~=0.19.4
python-dotenv~=0.21.0
Pillow~=9.2.0
appdirs~=1.4.4
future~=0.18.2
requests~=2.28.1
selenium~=4.5.0
pycryptodome~=3.15.0
gTTS~=2.2.4